# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AuthOptions": {
        "additionalProperties": false,
        "description": "Public sign-in options for the web UI (no credentials needed).",
        "properties": {
          "demo_enabled": {
            "title": "Demo Enabled",
            "type": "boolean"
          },
          "environment": {
            "$ref": "#/components/schemas/DeploymentEnvironment"
          }
        },
        "required": [
          "environment",
          "demo_enabled"
        ],
        "title": "AuthOptions",
        "type": "object"
      },
...
      "DeploymentEnvironment": {
        "description": "Supported deployment tiers. Hosted multi-tenant deployment is absent until P07.\n\n``hosted`` is a single-user deployment behind an HTTPS reverse proxy (see\ndocs/DEPLOYMENT.md): public bind, Host-header allowlist, HTTPS-only web origins and secure\ncookies. It is not a multi-tenant or SSO deployment.",
        "enum": [
          "local",
          "test",
          "hosted"
        ],
        "title": "DeploymentEnvironment",
        "type": "string"
      },
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "description": "Stable machine-readable error code",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "description": "Human-readable explanation without secrets or host paths",
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MembershipRole": {
        "enum": [
          "owner",
          "admin",
          "member",
          "viewer"
        ],
        "title": "MembershipRole",
        "type": "string"
      },
...
      "PrincipalResponse": {
        "additionalProperties": false,
        "properties": {
          "auth_method": {
            "title": "Auth Method",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "is_demo": {
            "description": "Signed in with the shared demo account (demo workspace only)",
            "title": "Is Demo",
            "type": "boolean"
          },
          "is_operator": {
            "title": "Is Operator",
            "type": "boolean"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          },
          "workspaces": {
            "items": {
              "$ref": "#/components/schemas/WorkspaceGrantResponse"
            },
            "title": "Workspaces",
            "type": "array"
          }
        },
        "required": [
          "user_id",
          "subject",
          "display_name",
          "auth_method",
          "is_operator",
          "is_demo",
          "workspaces"
        ],
        "title": "PrincipalResponse",
        "type": "object"
      },
...
      "SessionRequest": {
        "additionalProperties": false,
        "properties": {
          "token": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token"
        ],
        "title": "SessionRequest",
        "type": "object"
      },
      "SessionResponse": {
        "additionalProperties": false,
        "properties": {
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          }
        },
        "required": [
          "subject",
          "expires_at"
        ],
        "title": "SessionResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorkspaceGrantResponse": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/MembershipRole"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "workspace_id": {
            "format": "uuid",
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "workspace_id",
          "slug",
          "name",
          "role"
        ],
        "title": "WorkspaceGrantResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "refactorX control-plane API: authentication (access token or optional demo account), readiness, projects and sample projects, ZIP/local-runner intake, frozen snapshots, scans (PMD, ESLint, Opengrep, Trivy, structure, graph) with coverage, findings, issues, comparison and exports. Source-only analysis; no AI provider is used.",
    "title": "refactorX API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/auth/demo-session": {
      "post": {
        "description": "Sign in to the shared demo account (demo workspace only) when the demo is enabled.",
        "operationId": "create_demo_session_v1_auth_demo_session_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "summary": "Create Demo Session",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/me": {
      "get": {
        "operationId": "current_principal_v1_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrincipalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "summary": "Current Principal",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/options": {
      "get": {
        "description": "Public: which sign-in methods the web UI should offer (no credentials required).",
        "operationId": "sign_in_options_v1_auth_options_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthOptions"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Host not allowed (hosted mode)"
          }
        },
        "summary": "Sign In Options",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/session": {
      "delete": {
        "operationId": "delete_session_v1_auth_session_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "summary": "Delete Session",
        "tags": [
          "auth"
        ]
      },
      "post": {
        "description": "Exchange the local development token for an HttpOnly session cookie.",
        "operationId": "create_session_v1_auth_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Create Session",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}